make code more modular, fixes #11
