+++ src/edit_python_pe/main.py
@@ -1,22 +1,15 @@
-import getpass
 import glob
-import hashlib
 import os
-import pathlib
-import re
-from datetime import date, datetime
-from time import sleep
 
-import pygit2
-from github import Github
-from github.GithubException import BadCredentialsException, GithubException
 from github.Repository import Repository
-from platformdirs import user_data_dir
 from textual.app import App, ComposeResult
 from textual.containers import Horizontal, Vertical
 from textual.events import Event
 from textual.widgets import (Button, Input, ListItem, ListView, Select, Static,
                              TextArea)
+
+from .utils import (build_md_content, create_pr, fork_repo, get_repo,
+                    load_file_into_form)
 
 
 class MemberApp(App):
@@ -178,106 +171,8 @@ class MemberApp(App):
         self.current_file = filename
 
         self.clear_form()
-        self.load_file_into_form(filename)
+        load_file_into_form(self, filename)
         self.show_form()
-
-    def load_file_into_form(self, filename: str) -> None:
-        path_md = os.path.join(self.repo_path, "blog", "members", filename)
-        if not os.path.exists(path_md):
-            return
-        try:
-            with open(path_md, "r", encoding="utf-8") as fd:
-                content = fd.read()
-        except Exception as e:
-            self.exit(f"Error al leer el archivo {filename}: {e}")
-            return
-
-        self.clear_form()
-
-        # Extract YAML frontmatter
-        yaml_match = re.search(r"---\n(.*?)---\n", content, re.DOTALL)
-        yaml_data = {}
-        if yaml_match:
-            try:
-                import yaml
-
-                yaml_data = yaml.safe_load(yaml_match.group(1))
-            except Exception:
-                yaml_data = {}
-        self.name_input.value = yaml_data.get("author", "")
-        self.city_input.value = yaml_data.get("location", "")
-
-        # Extract member name from first markdown header
-        name_match = re.search(r"^# (.+)", content, re.MULTILINE)
-        if name_match:
-            self.name_input.value = name_match.group(1).strip()
-
-        # Extract gravatar email
-        gravatar_match = re.search(r"```\{gravatar\} ([^\n]+)", content)
-        if gravatar_match:
-            self.email_input.value = gravatar_match.group(1).strip()
-
-        # Extract social links from raw HTML block
-        social_block = re.search(
-            r"```\{raw\} html\n(.*?)```", content, re.DOTALL
-        )
-        if social_block:
-            social_html = social_block.group(1)
-            social_link_pattern = re.compile(
-                r'<a[^>]*href="([^"]+)"[^>]*>\s*<iconify-icon[^>]*icon="simple-icons:([^"]+)"',
-                re.DOTALL,
-            )
-            for match in social_link_pattern.finditer(social_html):
-                url = match.group(1)
-                platform = match.group(2)
-                self.add_social_entry()
-                last_entry = self.social_entries[-1]
-                last_entry.select.value = platform
-                last_entry.url_input.value = url
-
-        # Extract aliases, city, homepage from colon-prefixed lines
-        alias_match = re.search(r":Aliases:\s*([^\n]+)", content)
-        if alias_match:
-            aliases = [a.strip() for a in alias_match.group(1).split(",")]
-            for alias_val in aliases:
-                self.add_alias_entry()
-                self.alias_entries[-1].alias_input.value = alias_val
-        city_match = re.search(r":Ciudad:\s*([^\n]+)", content)
-        if city_match:
-            self.city_input.value = city_match.group(1).strip()
-        homepage_match = re.search(r":Homepage:\s*([^\n]+)", content)
-        if homepage_match:
-            self.homepage_input.value = homepage_match.group(1).strip()
-
-        # Extract markdown sections under headers
-        who_match = re.search(
-            r"### ¿Quién eres y a qué te dedicas\?\n(.*?)(?=^### |\Z)",
-            content,
-            re.DOTALL | re.MULTILINE,
-        )
-        if who_match:
-            self.who_area.text = who_match.group(1).strip()
-        python_match = re.search(
-            r"### ¿Cómo programas en Python\?\n(.*?)(?=^### |\Z)",
-            content,
-            re.DOTALL | re.MULTILINE,
-        )
-        if python_match:
-            self.python_area.text = python_match.group(1).strip()
-        contrib_match = re.search(
-            r"### ¿Tienes algún aporte a la comunidad de Python\?\n(.*?)(?=^### |\Z)",
-            content,
-            re.DOTALL | re.MULTILINE,
-        )
-        if contrib_match:
-            self.contributions_area.text = contrib_match.group(1).strip()
-        avail_match = re.search(
-            r"### ¿Estás disponible para hacer mentoring, consultorías, charlas\?\n(.*?)(?=^### |\Z)",
-            content,
-            re.DOTALL | re.MULTILINE,
-        )
-        if avail_match:
-            self.availability_area.text = avail_match.group(1).strip()
 
     def on_button_pressed(self, event: Button.Pressed) -> None:
         bid = event.button.id
@@ -421,89 +316,18 @@ class MemberApp(App):
                 socials.append((plat, urlval))
 
         # Build the markdown doc as per the provided guide
-        md_lines = [
-            "---",
-            "blogpost: true",
-            f"date: {date.today().strftime("%d %b, %Y")}",
-            f"author: {get_alias(aliases, name)}",
-            f"location: {city}",
-            "category: members",
-            "language: Español",
-            "image: 1",
-            "excerpt: 1",
-            "---",
-            "",
-            f"# {name}",
-            "",
-            f"```{{gravatar}} {email}",
-            "---",
-            "width: 200",
-            'class: "member-gravatar"',
-            "---",
-            "```",
-            "",
-        ]
-        if socials:
-            md_lines.append("```{raw} html")
-            md_lines.append('<ul class="social-media profile">')
-            for plat, url in socials:
-                md_lines.append("    <li>")
-                md_lines.append(
-                    f'        <a class="external reference" href="{url}">'
-                )
-                md_lines.append(
-                    f'            <iconify-icon icon="simple-icons:{plat}" style="font-size:2em"></iconify-icon>'
-                )
-                md_lines.append("        </a>")
-                md_lines.append("    </li>")
-            md_lines.append("</ul>")
-            md_lines.append("```")
-            md_lines.append("")
-
-        if aliases:
-            md_lines.append(f":Aliases: {', '.join(aliases)}")
-            md_lines.append("")
-
-        if city:
-            md_lines.append(f":Ciudad: {city}")
-            md_lines.append("")
-
-        if homepage:
-            md_lines.append(f":Homepage: {homepage}")
-            md_lines.append("")
-
-        md_lines.append("## Sobre mí")
-        md_lines.append("")
-
-        if who:
-            md_lines.append("### ¿Quién eres y a qué te dedicas?")
-            md_lines.append("")
-            md_lines.append(who)
-            md_lines.append("")
-
-        if python_:
-            md_lines.append("### ¿Cómo programas en Python?")
-            md_lines.append("")
-            md_lines.append(python_)
-            md_lines.append("")
-
-        if contributions:
-            md_lines.append(
-                "### ¿Tienes algún aporte a la comunidad de Python?"
-            )
-            md_lines.append("")
-            md_lines.append(contributions)
-            md_lines.append("")
-
-        if availability:
-            md_lines.append(
-                "### ¿Estás disponible para hacer mentoring, consultorías, charlas?"
-            )
-            md_lines.append("")
-            md_lines.append(availability)
-            md_lines.append("")
-
-        md_content = "\n".join(md_lines)
+        md_content = build_md_content(
+            name,
+            email,
+            aliases,
+            socials,
+            city,
+            homepage,
+            who,
+            python_,
+            contributions,
+            availability,
+        )
 
         message = create_pr(
             md_content,
@@ -524,214 +348,6 @@ class MemberApp(App):
             self.on_list_view_selected(event)
 
         await super().on_event(event)
-
-
-def get_repo() -> tuple[str, Repository]:
-    token = getpass.getpass(
-        "Por favor ingrese su access token personal de GitHub: "
-    )
-    g = Github(token)
-
-    try:
-        return token, g.get_repo("pythonpe/python.pe")
-    except BadCredentialsException:
-        print("Acceso no autorizado. Por favor, verifique su token de acceso.")
-        exit(1)
-    except GithubException:
-        print(
-            "Repositorio no encontrado. Por favor, verifique su token de acceso."
-        )
-        exit(1)
-
-
-def fork_repo(token: str, original_repo: Repository) -> tuple[str, Repository]:
-    forked_repo = original_repo.create_fork()
-    forked_repo_url = forked_repo.clone_url
-    repo_path = user_data_dir(appname="edit-python-pe", appauthor="python.pe")
-
-    if not os.path.exists(repo_path):
-        callbacks = pygit2.callbacks.RemoteCallbacks(
-            credentials=pygit2.UserPass(token, "x-oauth-basic")
-        )
-        sleep(3)
-        pygit2.clone_repository(
-            forked_repo_url, repo_path, callbacks=callbacks
-        )
-    return repo_path, forked_repo
-
-
-def compute_file_name(aliases: list[str], name: str, email: str) -> str:
-    # compute name_file
-    if aliases:
-        alias_for_name = aliases[0].lower().replace(" ", "_")
-    else:
-        alias_for_name = name.lower().replace(" ", "_")
-
-    sha_hash = hashlib.sha1(
-        (alias_for_name + email + datetime.now().isoformat()).encode("utf-8")
-    ).hexdigest()[:8]
-    return f"{alias_for_name}-{sha_hash}.md"
-
-
-def read_file(file_path: str) -> str:
-    with open(file_path, "r", encoding="utf-8") as fd:
-        return fd.read()
-
-
-def append_file(file_content: str, file_path: str) -> None:
-    os.makedirs(os.path.dirname(file_path), exist_ok=True)
-    with open(file_path, "a", encoding="utf-8") as fd:
-        fd.write(file_content)
-
-
-def write_file(file_content: str, file_path: str) -> None:
-    os.makedirs(os.path.dirname(file_path), exist_ok=True)
-    with open(file_path, "w", encoding="utf-8") as fd:
-        fd.write(file_content)
-
-
-def commit_and_push(
-    repo_path: str,
-    token: str,
-    was_changed: bool,
-    name_file: str,
-    name: str,
-    email: str,
-) -> tuple[
-    str,
-    pygit2.repository.Repository,
-    pygit2.remotes.Remote,
-    pygit2.callbacks.RemoteCallbacks,
-]:
-    repo = pygit2.repository.Repository(repo_path)
-    repo.index.add_all()
-    repo.index.write()
-    author_sig = pygit2.Signature(name or "Unknown", email or "unknown@email")
-    tree_id = repo.index.write_tree()
-    parents = [] if repo.head_is_unborn else [repo.head.target]
-    commit_msg = (
-        f"Changed {name_file}" if was_changed else f"Added {name_file}"
-    )
-    repo.create_commit(
-        "HEAD", author_sig, author_sig, commit_msg, tree_id, parents
-    )
-
-    callbacks = pygit2.callbacks.RemoteCallbacks(
-        credentials=pygit2.UserPass(token, "x-oauth-basic")
-    )
-    remote = repo.remotes["origin"]
-    remote.push([repo.head.name], callbacks=callbacks)
-    return commit_msg, repo, remote, callbacks
-
-
-def get_alias(aliases: list[str], name: str) -> str:
-    if aliases:
-        return aliases[0]
-    return name
-
-
-def create_member_file(
-    file_content: str,
-    current_file: str | None,
-    repo_path: str,
-    aliases: list[str],
-    name: str,
-    email: str,
-) -> tuple[str, str]:
-    # Name the file
-    name_file = (
-        current_file
-        if current_file is not None
-        else compute_file_name(aliases, name, email)
-    )
-
-    # Write file
-    file_path = os.path.join(repo_path, "blog", "members", name_file)
-    write_file(file_content, file_path)
-    return name_file, file_path
-
-
-def write_authors_file(
-    repo_path: str, aliases: list[str], name: str, email: str
-):
-    file_path = os.path.join(repo_path, "AUTHORS")
-    contents = read_file(file_path)
-    file_content = f"\n{name}({get_alias(aliases, name)}) <{email}>"
-    if file_content.strip() not in contents:
-        append_file(file_content, file_path)
-
-
-def create_pr(
-    file_content: str,
-    current_file: str | None,
-    repo_path: str,
-    original_repo: Repository,
-    forked_repo: Repository,
-    token: str,
-    aliases: list[str],
-    name: str,
-    email: str,
-) -> str:
-    name_file, file_path = create_member_file(
-        file_content, current_file, repo_path, aliases, name, email
-    )
-    write_authors_file(repo_path, aliases, name, email)
-
-    # commit & push
-    commit_msg, repo, remote, callbacks = commit_and_push(
-        repo_path,
-        token,
-        current_file is not None,
-        name_file,
-        name,
-        email,
-    )
-
-    # PR logic
-    pr_title = commit_msg
-    first_alias = aliases[0] if aliases else ""
-    pr_body = (
-        f"Changing an entry to `blog/members` for {name} (alias: {first_alias})."
-        if current_file
-        else f"Creating a new entry to `blog/members` for {name} (alias: {first_alias})."
-    )
-    fork_owner = forked_repo.owner.login
-    head_branch = f"{fork_owner}:main"
-    base_branch = "main"
-
-    # If editing, retrieve PR by title and push to its branch
-    if current_file:
-        # Try to find an open PR with matching title
-        prs = original_repo.get_pulls(
-            state="open", sort="created", base=base_branch
-        )
-        pr_found = None
-        for pr in prs:
-            if current_file in pr.title:
-                pr_found = pr
-                break
-        if pr_found:
-            # Push to the PR branch (simulate, as actual branch logic may differ)
-            remote.push([repo.head.name], callbacks=callbacks)
-            return f"Archivo {name_file} editado, commit y cambios enviados al PR existente."
-        else:
-            # Otherwise, create a new PR
-            original_repo.create_pull(
-                title=pr_title,
-                body=pr_body,
-                head=head_branch,
-                base=base_branch,
-            )
-            return f"Archivo {name_file} guardado, commit y PR listo."
-    else:
-        # Otherwise, create a new PR
-        original_repo.create_pull(
-            title=pr_title,
-            body=pr_body,
-            head=head_branch,
-            base=base_branch,
-        )
-        return f"Archivo {name_file} guardado, commit y PR listo."
 
 
 def main() -> None:
